+++ src/Simulate.jl
@@ -43,15 +43,15 @@ import Unitful: FreeUnits, Time
 
 include("types.jl")
 include("components.jl")
-include("events.jl")
 include("schedule.jl")
+include("events.jl")
 include("clock.jl")
 include("process.jl")
 include("threads.jl")
 include("timer.jl")
 include("printing.jl")
 
-export  Clock, ActiveClock, RTClock, RTC, setUnit!, 𝐶, 
+export  Clock, ActiveClock, RTClock, RTC, setUnit!, 𝐶,
         Timing, at, after, every, before, until, tau, sample_time!,
         Action, Fun, event!, periodic!,
         incr!, run!, stop!, resume!, reset!, sync!,